+++ benchopt/tests/test_runner.py
@@ -46,26 +46,14 @@ def test_skip_api():
     assert reason == 'X is all zeros'
 
 
-def test_get_one_solution():
+def test_get_one_result():
     dataset = TEST_DATASET.get_instance()
     objective = TEST_OBJECTIVE.get_instance()
     objective.set_dataset(dataset)
 
-    one_solution = objective.get_one_solution()
+    one_solution = objective.get_one_result()
     expected = np.zeros(objective.X.shape[1])
-    assert all(one_solution == expected)
-
-    # XXX - Remove in version 1.3
-    dataset = TEST_DATASET.get_instance(deprecated_return=True)
-    objective = TEST_OBJECTIVE.get_instance(deprecated_dataset=True)
-
-    with pytest.warns(FutureWarning, match="`get_data` should return a dict"):
-        objective.set_dataset(dataset)
-
-    with pytest.warns(FutureWarning, match="Objective should have a method"):
-        one_solution = objective.get_one_solution()
-    expected = np.zeros(objective.X.shape[1])
-    assert all(one_solution == expected)
+    assert all(one_solution['beta'] == expected)
 
 
 def _assert_parameters_equal(instance, parameters):