+++ benchopt/utils/parametrized_name_mixin.py
@@ -38,9 +38,20 @@ class ParametrizedNameMixin():
         try:
             obj = cls(**parameters)
             obj.save_parameters(**parameters)
-        except Exception:
-            print(cls)
+        except Exception as exception:
+            # get type (Dataset, Objective, or Solver) of class and its name
+            cls_type = cls.__base__.__name__
+            cls_type = cls_type.replace("Base", "")
+            cls_name = cls.name
+
+            # Extend exception error message
+            # TODO: use `add_note` when requiring python>=3.11
+            exception.args = (
+                f'Error when initializing {cls_type}: "{cls_name}". '
+                f'{". ".join(exception.args)}',
+            )
             raise
+
         return obj
 
     @property